Début du développement des méthodes de Account, petite correction sur current account

+++ Current_Account.py
@@ -4,7 +4,7 @@ from Account import Account
 class Current_Account(Account):
     """"Class inheriting from the Account class"""
     overdraft: int
-    agios_percentage: int
+    agios_percentage: float
 
 
     def apply_agios(self):
